Added local thumbnail generation via chrome.tabs.captureVisibleTab() in out eventPage.js, and make sure the captured data doesn't get synchronized since it's too big to sync and not worth syncing.
Made sure to only parse this data structure ONCE and keep it in memory to reference, re-parsing when needed is too big a performance hit to do it every time we need the data.

Had to add permissions to allow taking local screen shots, unfortunately this can't be done without asking for both "tabs" and <all_urls>. "tabs" is needed to get the url of the tab since "activeTab" isn't allowed seeing the tab's url, and <all_tabs> is required by the chrome.tabs.captureVisibleTab() API call which is noted here: https://developer.chrome.com/extensions/tabs#method-captureVisibleTab

Changed entry dimensions to 16:10 instead of 4:3 to adapt to taking local screen shot data and altered favicon sizes accordingly.

Yes this is a big change... but had to happen sometime.